pythonProject/hf_space/requirements.txt
Flask==2.3.2
sentence-transformers==2.2.2
vncorenlp==1.0.3